+++ regix/__init__.py
@@ -129,4 +129,4 @@ __all__ = [
     "ArchSmell",
 ]
 
-__version__ = "0.1.25"
+__version__ = "0.1.26"